+++ app/src/main/java/dk/eleknet/atool/extensions/Boolean.kt
@@ -1,0 +1,5 @@
+package dk.eleknet.atool.extensions
+
+fun Boolean?.orValue(value: Boolean): Boolean {
+    return this ?: value
+}